Tested up until ASM generation for [1, 16] go routine parallelism

+++ src/frontend/tree.go
@@ -82,7 +82,7 @@ func nodeInit(typ ir.NodeType, data interface{}, line, pos int, args ...yySymTyp
 		}
 	case ir.FLOAT_DATA:
 		if num, err := parseFloat(data); err == nil {
-			n.Data = float32(num)
+			n.Data = float64(num)
 		} else {
 			fmt.Println(err)
 			n.Data = data.(string)
